% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_ecg.R
\name{read_ecg}
\alias{read_ecg}
\title{Read ECG data file}
\usage{
read_ecg(NAME = "filename")
}
\arguments{
\item{NAME}{The input parameter is the string path to the selected ECG file (csv format).}
}
\value{
Function returns the ecg data frame.
}
\description{
This function reads raw ECG data acquired by Zephyr BH3 and sorts it into a data frame.
The stringsAsFactors parameter is set to FALSE by default.
}
\examples{
data_file <- system.file("extdata", "myECGData.csv", package="zephyrECG")
ecg <- read_ecg(data_file)
}
\keyword{ECG}
\keyword{data,}
\keyword{read}

