% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/separate_bh3.R
\name{separate_bh3}
\alias{separate_bh3}
\title{Zephyr BioHarness 3 data file separation}
\usage{
separate_bh3(NAME = "filename")
}
\arguments{
\item{NAME}{Input for the function is a string with the path to the BH3 file.}
}
\value{
Function doesn't return any objects to the environment. Instead it writes the separated sessions as csv files to the same directory as the source data file.
}
\description{
This function separates ECG sessions from the Zephyr BioHarness 3 file to single csv files.
Zephyr BH3 unix timestamps are converted to date-time format Y-m-d HH:MM:SS.ms(3 digits).
The function writes the separate ECG sessions as csv files to the same directory.
}
\examples{
data_file <- system.file("extdata", "myZephyrBH3Data.csv", package="zephyrECG")
separate_bh3(data_file)
}
\keyword{BH3}
\keyword{Zephyr}
\keyword{data}
\keyword{file,}
\keyword{separate}

